Refactor target type and technology stack enums into shared module
