+++ shared/target_types.py
@@ -1,0 +1,29 @@
+from enum import Enum
+
+class TargetType(Enum):
+    """Enumeration of different target types for intelligent analysis"""
+    WEB_APPLICATION = "web_application"
+    NETWORK_HOST = "network_host"
+    API_ENDPOINT = "api_endpoint"
+    CLOUD_SERVICE = "cloud_service"
+    MOBILE_APP = "mobile_app"
+    BINARY_FILE = "binary_file"
+    UNKNOWN = "unknown"
+
+class TechnologyStack(Enum):
+    """Common technology stacks for targeted testing"""
+    APACHE = "apache"
+    NGINX = "nginx"
+    IIS = "iis"
+    NODEJS = "nodejs"
+    PHP = "php"
+    PYTHON = "python"
+    JAVA = "java"
+    DOTNET = "dotnet"
+    WORDPRESS = "wordpress"
+    DRUPAL = "drupal"
+    JOOMLA = "joomla"
+    REACT = "react"
+    ANGULAR = "angular"
+    VUE = "vue"
+    UNKNOWN = "unknown"
